+++ src/main/java/com/tallerwebi/dominio/ServicioSubasta.java
@@ -1,5 +1,5 @@
 package com.tallerwebi.dominio;
 
 public interface ServicioSubasta {
-    void crearSubasta(Subasta subasta);
+    void crearSubasta(Subasta subasta, String creador);
 }
