+++ src/main/java/com/tallerwebi/presentacion/ControladorLogin.java
@@ -19,16 +19,11 @@ import javax.servlet.http.HttpServletRequest;
 public class ControladorLogin {
 
     private ServicioLogin servicioLogin;
-    private ServicioSubasta servicioSubasta;
 
     @Autowired
-    public ControladorLogin(ServicioLogin servicioLogin, ServicioSubasta servicioSubasta) {
+    public ControladorLogin(ServicioLogin servicioLogin) {
         this.servicioLogin = servicioLogin;
-        this.servicioSubasta = servicioSubasta;
     }
-
-    //NOTA: Creo que el profe se enojara si pongo 2 servicios en 1 controlador. Despues arreglare eso.
-
 
 
     @RequestMapping("/login")
@@ -45,7 +40,6 @@ public class ControladorLogin {
         Usuario usuarioBuscado = servicioLogin.consultarUsuario(datosLogin.getEmail(), datosLogin.getPassword());
         if (usuarioBuscado != null) {
             request.getSession().setAttribute("USUARIO", usuarioBuscado.getEmail());
-            request.getSession().setAttribute("PASS", usuarioBuscado.getPassword());
             request.getSession().setAttribute("ROL", usuarioBuscado.getRol());
             return new ModelAndView("redirect:/home");
         } else {
@@ -86,30 +80,5 @@ public class ControladorLogin {
         return new ModelAndView("redirect:/login");
     }
 
-    //TO-DO: Mover estas clases de subastas a un controlador exclusivo, en vez de usar el de Login.
-    @RequestMapping(path = "/nuevaSubasta", method = RequestMethod.GET)
-    public ModelAndView irANuevaSubasta() {
-        ModelMap model = new ModelMap();
-        model.put("subasta", new Subasta());
-        return new ModelAndView("nuevaSubasta", model);
-    }
-
-    @RequestMapping(path = "/crearSubasta", method = RequestMethod.POST)
-    public ModelAndView crearSubasta(@ModelAttribute("subasta") Subasta subasta, HttpServletRequest request) {
-
-        ModelMap model = new ModelMap();
-        Usuario creador;
-        String creadorEmail = (String) request.getSession().getAttribute("USUARIO");
-        String creadorPass = (String) request.getSession().getAttribute("PASS");
-        creador = servicioLogin.consultarUsuario(creadorEmail, creadorPass);
-        subasta.setCreador(creador);
-        try{
-            servicioSubasta.crearSubasta(subasta);
-        }catch (Exception e){
-            model.put("error", "Hubo un error al crear la subasta. Intentelo nuevamente.");
-            return new ModelAndView("nuevaSubasta", model);
-        }
-        return new ModelAndView("redirect:/home");
-    }
 }
